requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
matplotlib>=3.4.0
seaborn>=0.11.0
paho-mqtt>=1.6.0
fastapi>=0.75.0
uvicorn>=0.17.0
websockets>=10.0
psutil>=5.8.0
memory-profiler>=0.60.0
joblib>=1.1.0
asyncio-mqtt>=0.11.0
pyyaml
mosquitto
pytest
pytest-cov